FORCE: Reestruturação do servidor
