--- public/models/Publicacao.js
@@ -1,8 +1,0 @@
-export class Publicacao {
-  _id = 0;
-  nome = "";
-  titulo = "";
-  depoimento = "";
-  dataHora = "";
-  __v = 0;
-}
