+++ server/index.js
@@ -1,9 +1,6 @@
 const logs = require('./logs')
 const Hapi = require('@hapi/hapi')
 const routes = require("./routes")
-// require('dotenv/config')
-// const { HAPI_PORT } = process.env;
-// const { HAPI_HOST } = process.env;
 
 const init = async () => {
   const server = new Hapi.Server({
@@ -23,6 +20,7 @@ await server.start()
 console.log(logs.printhash())
 console.log(`Server running at: ${server.info.uri}`)
 console.log(`Pressione ctrl + C para finalizar o servidor!`)
+console.log(logs.printhash())
 } // Fim da declaração da função init
 process.on('unhandledRejection', (err) => {
   console.error(err)
