+++ server/public/galeria.js
