+++ server/routes.js
@@ -3,18 +3,18 @@ const handles = require("./handles")
 module.exports = [
   {
     method: 'GET',
-    path: '/api/v1/publicacao',
+    path: '/api/v1/publicacoes',
     handler: handles.listarPublicacoes
   },
   {
     method: 'POST',
-    path: '/api/v1/publicacao',
+    path: '/api/v1/publicacoes',
     handler: handles.criarPublicacao
   },
   {
     method: 'GET',
     path: '/api/v1/usuarios',
-    handler: handles.listarUsuarios
+    handler: handles.listarUsuario
   },
   {
     method: 'POST',
